fxa-client: Remove BrowserID assertions, rename feature to "session_token".
This removes support for generating BrowserID assertions, but keeps some
of the other session-token-using machinery that was beind the "browserid"
feature flag. Since this functionality now has nothing to do with BrowserID
the feature flag has been renamed to "session_token".
